ENH: mark the dragon4 scratch space as thread-local
